+++ ucte/ucte-network/src/main/java/com/powsybl/ucte/network/UcteLine.java
@@ -14,7 +14,7 @@ import com.powsybl.commons.reporter.Reporter;
  */
 public class UcteLine extends UcteElement {
 
-    public UcteLine(UcteElementId id, UcteElementStatus status, float resistance, float reactance, float susceptance, Integer currentLimit, String elementName) {
+    public UcteLine(UcteElementId id, UcteElementStatus status, double resistance, double reactance, double susceptance, Integer currentLimit, String elementName) {
         super(id, status, resistance, reactance, susceptance, currentLimit, elementName);
 
         if (id.getNodeCode1().getVoltageLevelCode() != id.getNodeCode2().getVoltageLevelCode()) {
